requirements.txt
tensorflow
pillow
scipy
numpy
Flask

# Add other dependencies here

